Allow running of 'mvn test'
There was a dependency:unpack goal bound to the process-sources phase
which would fail if you ran 'mvn test' because the artifact had not been
packaged.  Since this artifact isn't actually needed until
integration-test, change the phase binding to pre-integration-test.

There were a couple of other goals bound to other phases that would be
better bound to pre-integration-test, so those are fixed as well.