Update answer store population instructions
Classifier training and canonical question mapping files are formatted
as csv files, but we still need a json version of the training file at
runtime, and we need to generate a single json file that can be POSTed
to populate the answer store.

Modify GenerateTrainingAndPopulationData to generate these files, and
call it during the build.

Update the instructions in the readme to reflect the current usage of
json and csv files.

Note that training.json has changed, but it's purely a re-ordering of
the entries.