Fix liberty:stop-server command
When liberty is running the Liberty-is-running profile is activated.
(So we can stop liberty during a clean.)  However, this causes the
Liberty profile to not be activated.

Move the install config up into the common plugin config, and change the
Liberty profile activation so it's active any time bluemix.appname isn't
set.